Web: mode global events constants from client to common
