+++ packages/client/src/store/DialogsStore.js
@@ -1,7 +1,7 @@
 import { getNewFiles } from "@docspace/common/api/files";
 import { FileAction } from "@docspace/common/constants";
 import { makeAutoObservable } from "mobx";
-import { Events } from "@docspace/client/src/helpers/filesConstants";
+import { Events } from "@docspace/common/constants";
 
 class DialogsStore {
   authStore;
